Don't highlight dragon cards during a drag
Most of the time, when dragging cards, the dragged card helper receives
mouse events instead of whatever is underneath those cards. But if the
mouse moves fast enough, it's possible to get the mouseenter event on a
dragon button while before the dragged cards catch up and cover it,
causing a distracting flicker on the matching dragon cards.

This adds a global that is set while a drag is in progress, allowing the
mouseover events to exit early and not produce the flicker.